+++ device-minion-runtime/src/WorkerMinion/include/trace.h
@@ -18,6 +18,13 @@
 
 #include <et-trace/encoder.h>
 
+/*! \def TRACE_STRING_MAX_SIZE_CM
+    \brief Max string message lentgh which can be logged into Trace.
+        NOTE: This will be removed as a result of SW-13550. Because
+        it will direclty use Trace encoder for string formatting.
+*/
+#define TRACE_STRING_MAX_SIZE_CM 128
+
 /*! \fn void Trace_Init_CM(const struct trace_init_info_t *cm_init_info)
     \brief This function initializes Trace for a single Hart in CM Shires.
            All CM Harts must call this function to Enable Trace.